HV:common:fix "integer type violations"
fix integer type violations,keep some violations which
related to hypcall and msix_entry_index.

V1->V2:1.modified API_MAJOR_VERSION from Makefile
       2.sync acrn_common.h changed to device model

Acked-by: Eddie Dong <eddie.dong@intel.com>

+++ hypervisor/common/hv_main.c
@@ -21,7 +21,7 @@ static void run_vcpu_pre_work(struct vcpu *vcpu)
 void vcpu_thread(struct vcpu *vcpu)
 {
 	uint64_t vmexit_begin = 0UL, vmexit_end = 0UL;
-	uint16_t basic_exit_reason = 0U;
+	uint32_t basic_exit_reason = 0U;
 	uint64_t tsc_aux_hyp_cpu = (uint64_t) vcpu->pcpu_id;
 	int32_t ret = 0;
